requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.24.1
pandas==1.3.3
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.28.1
pytz==2022.1
schedule==1.1.0
python-telegram-bot==13.7
pydantic==1.8.2
tenacity==8.2.2
pytest==7.4.2
financedatareader==0.9.50 